Update All tanggal 11 Maret 2015

+++ osipkd/views/akuntansi/ak_jurnal.py
@@ -21,11 +21,21 @@ def deferred_jv_type(node, kw):
     return widget.SelectWidget(values=values)
     
 JV_TYPE = (
-    (1, 'LRA'),
-    (2, 'LO'),
+    (1, 'Jurnal Penerimaan'),
+    (2, 'Jurnal Pengeluaran'),
     (3, 'Jurnal Umum'),
+    (4, 'Jurnal Koreksi'),
+    (5, 'Jurnal Penutup'),
     )
 
+def deferred_is_skpd(node, kw):
+    values = kw.get('is_skpd', [])
+    return widget.SelectWidget(values=values)
+    
+IS_SKPD = (
+    (0, 'PPKD'),
+    (1, 'SKPD'))   
+    
 class AddSchema(colander.Schema):
     unit_kd_widget = widget.AutocompleteInputWidget(
             values = '/unit/act/headofkode',
@@ -36,6 +46,11 @@ class AddSchema(colander.Schema):
             min_length=1)
             
           
+    tahun_id  = colander.SchemaNode(
+                    colander.Integer(),
+                    oid='tahun_id',
+                    title="Tahun")
+
     unit_id  = colander.SchemaNode(
                     colander.Integer(),
                     oid='unit_id',
@@ -54,10 +69,13 @@ class AddSchema(colander.Schema):
                     widget = unit_nm_widget)
     kode        = colander.SchemaNode(
                     colander.String(),
+                    missing = colander.drop,
+                    title="No. Jurnal"
                     )
                     
     nama        = colander.SchemaNode(
                     colander.String(),
+                    title="Uraian"
                     )
     tanggal     = colander.SchemaNode(
                   colander.Date(),
@@ -79,7 +97,7 @@ class AddSchema(colander.Schema):
                     )
     tgl_source  = colander.SchemaNode(
                     colander.Date(),
-                    title = "Tgl"
+                    title = "Tgl. Ref"
                 )
     
     notes       = colander.SchemaNode(
@@ -87,7 +105,10 @@ class AddSchema(colander.Schema):
                     missing = colander.drop
                     )
     is_skpd     = colander.SchemaNode(
-                    colander.Boolean()
+                    colander.Integer(),
+                    title="Jurnal",
+                    oid = "is_skpd",
+                    widget=widget.SelectWidget(values=IS_SKPD)
                   )
                     
                     
@@ -125,16 +146,18 @@ class view_ak_jurnal(BaseViews):
             columns.append(ColumnDT('id'))
             columns.append(ColumnDT('tanggal', filter=self._DTstrftime))
             columns.append(ColumnDT('kode'))
+            columns.append(ColumnDT('jv_type'))
             columns.append(ColumnDT('nama'))
+            columns.append(ColumnDT('source'))
             columns.append(ColumnDT('amount',  filter=self._number_format))
             columns.append(ColumnDT('posted'))
             
-            query = DBSession.query(Jurnal.id, Jurnal.tanggal, Jurnal.kode, 
-                      Jurnal.nama, Jurnal.posted,
-                      func.div(func.coalesce(func.sum(JurnalItem.amount),0),2).label('amount')).\
+            query = DBSession.query(Jurnal.id, Jurnal.tanggal, Jurnal.kode, Jurnal.jv_type,
+                      Jurnal.nama, Jurnal.source, Jurnal.posted,
+                      func.coalesce(func.sum(JurnalItem.amount),0).label('amount')).\
                     outerjoin(JurnalItem).\
-                    group_by(Jurnal.id, Jurnal.tanggal, Jurnal.kode, 
-                             Jurnal.nama, ).\
+                    group_by(Jurnal.id, Jurnal.tanggal, Jurnal.kode, Jurnal.jv_type,
+                             Jurnal.nama, Jurnal.source, ).\
                     filter(Jurnal.tahun_id == ses['tahun'],
                            Jurnal.unit_id == ses['unit_id'],)
                       
@@ -154,7 +177,7 @@ class view_ak_jurnal(BaseViews):
                 
     def get_form(self, class_form, row=None):
         schema = class_form(validator=self.form_validator)
-        schema = schema.bind(ap_type=AP_TYPE)
+        schema = schema.bind(jv_type=JV_TYPE, is_skpd=IS_SKPD)
         schema.request = self.request
         if row:
           schema.deserialize(row)
@@ -166,16 +189,26 @@ class view_ak_jurnal(BaseViews):
             row.created = datetime.now()
             row.create_uid = user.id
         row.from_dict(values)
-        tanggal    = datetime.strptime(values['tanggal'], '%Y-%m-%d')
-        row.tahun_id  = tanggal.year
-        row.periode  = tanggal.month
-        row.hari   = tanggal.day
+        tanggal           = datetime.strptime(values['tanggal'], '%Y-%m-%d')
+        row.tahun_id      = tanggal.year
+        row.periode       = tanggal.month
+        row.updated       = datetime.now()
+        row.update_uid    = user.id
+        row.disabled      = 'disabled' in values and values['disabled'] and 1 or 0
+        row.posted        = 'posted'  in values and values['posted']  and 1 or 0
+        row.tgl_transaksi = datetime.now()
         
-        row.updated = datetime.now()
-        row.update_uid = user.id
-        row.disable   = 'disable' in values and values['disable'] and 1 or 0
-        row.is_skpd   = 'is_skpd' in values and values['is_skpd'] and 1 or 0
-        row.posted    = 'posted' in values and values['posted'] and 1 or 0
+        if not row.kode:
+            tahun    = self.session['tahun']
+            unit_kd  = self.session['unit_kd']
+            is_skpd  = row.is_skpd
+            jv_type  = row.jv_type
+            tipe     = Jurnal.get_tipe(jv_type)
+            no_urut  = Jurnal.get_norut(row.id)+1
+            no       = "0000%d" % no_urut
+            nomor    = no[-5:]     
+            row.kode = "%d" % tahun + "-%s" % is_skpd + "-%s" % unit_kd + "-%s" % tipe + "-%s" % nomor
+        
         DBSession.add(row)
         DBSession.flush()
         return row
@@ -187,16 +220,11 @@ class view_ak_jurnal(BaseViews):
         self.request.session.flash('Jurnal sudah disimpan.')
         return row
             
-    def route_list(self, id):
-        if id:
-            return HTTPFound(location=self.request.route_url('ak-jurnal-edit', id=id) )
-        else:
-            return HTTPFound(location=self.request.route_url('ak-jurnal', id=id) )
+    def route_list(self):
+        return HTTPFound(location=self.request.route_url('ak-jurnal'))
         
     def session_failed(self, session_name):
-        #r = dict(form=self.session[session_name])
         del self.session[session_name]
-        #return r
         
     @view_config(route_name='ak-jurnal-add', renderer='templates/ak-jurnal/add.pt',
                  permission='add')
@@ -210,17 +238,11 @@ class view_ak_jurnal(BaseViews):
                 try:
                     c = form.validate(controls)
                 except ValidationFailure, e:
-                    #req.session[SESS_ADD_FAILED] = e.render()     
-                    #form.set_appstruct(rowd)
                     return dict(form=form)
-                    #return HTTPFound(location=req.route_url('ak-jurnal-add'))
-                id = self.save_request(dict(controls)).id
-            return self.route_list(id)
-            
+                id = self.save_request(dict(controls))
+            return self.route_list()  
         elif SESS_ADD_FAILED in req.session:
             return dict(form=form)
-        
-            #return self.session_failed(SESS_ADD_FAILED)
         return dict(form=form)
 
         
@@ -239,10 +261,14 @@ class view_ak_jurnal(BaseViews):
                  permission='edit')
     def view_ak_jurnal_edit(self):
         request = self.request
-        row = self.query_id().first()
+        row     = self.query_id().first()
+        
         if not row:
             return id_not_found(request)
-        #values = row.to_dict()
+        if row.posted:
+            request.session.flash('Data sudah diposting', 'error')
+            return self.route_list()
+            
         rowd={}
         rowd['id']            = row.id
         rowd['unit_id']       = row.unit_id
@@ -256,6 +282,8 @@ class view_ak_jurnal(BaseViews):
         rowd['tanggal']       = row.tanggal
         rowd['jv_type']       = row.jv_type
         rowd['disabled']      = row.disabled
+        rowd['notes']         = row.notes
+        rowd['is_skpd']       = row.is_skpd
         
         form = self.get_form(EditSchema)
         form.set_appstruct(rowd)
@@ -267,11 +295,8 @@ class view_ak_jurnal(BaseViews):
                     c = form.validate(controls)
                 except ValidationFailure, e:
                     return dict(form=form)
-                    #request.session[SESS_EDIT_FAILED] = e.render()               
-                    #return HTTPFound(location=request.route_url('ak-jurnal-edit',
-                    #                  id=row.id))
-                id = self.save_request(dict(controls)).id
-            return self.route_list(id)
+                self.save_request(dict(controls),row)
+            return self.route_list()
         elif SESS_EDIT_FAILED in request.session:
             return self.session_failed(SESS_EDIT_FAILED)
         return dict(form=form)
@@ -283,22 +308,25 @@ class view_ak_jurnal(BaseViews):
                  permission='delete')
     def view_ak_jurnal_delete(self):
         request = self.request
-        q = self.query_id()
-        row = q.first()
+        q       = self.query_id()
+        row     = q.first()
         
         if not row:
             return self.id_not_found(request)
+        
+        i = DBSession.query(JurnalItem).filter(JurnalItem.jurnal_id==row.id).first()         
+        if i:
+            request.session.flash('Hapus data item dahulu', 'error')
+            return self.route_list()
+            
         form = Form(colander.Schema(), buttons=('hapus','batal'))
         if request.POST:
             if 'hapus' in request.POST:
-                msg = 'Jurnal ID %d %s sudah dihapus.' % (row.id, row.nama)
-                try:
-                  q.delete()
-                  DBSession.flush()
-                except:
-                  msg = 'Jurnal ID %d %s tidak dapat dihapus.' % (row.id, row.nama)
-                request.session.flash(msg)
+                #Untuk hapus jurnal 
+                msg = '%s dengan kode %s telah berhasil.' % (request.title, row.kode)
+                q.delete()
+                DBSession.flush()
+                request.session.flash(msg)  
             return self.route_list()
-        return dict(row=row,
-                     form=form.render())
+        return dict(row=row, form=form.render())
